+++ pilot/pkg/networking/core/v1alpha3/cluster_test.go
@@ -25,7 +25,6 @@ import (
 
 	cluster "github.com/envoyproxy/go-control-plane/envoy/config/cluster/v3"
 	core "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
-	tls "github.com/envoyproxy/go-control-plane/envoy/extensions/transport_sockets/tls/v3"
 	"github.com/gogo/protobuf/proto"
 	"github.com/gogo/protobuf/types"
 	"github.com/golang/protobuf/ptypes"
@@ -43,7 +42,6 @@ import (
 	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pilot/pkg/networking/plugin"
 	"istio.io/istio/pilot/pkg/networking/util"
-	authn_model "istio.io/istio/pilot/pkg/security/model"
 	"istio.io/istio/pilot/test/xdstest"
 	"istio.io/istio/pkg/config"
 	"istio.io/istio/pkg/config/constants"
@@ -1869,1139 +1867,6 @@ func TestApplyLoadBalancer(t *testing.T) {
 	}
 }
 
-func TestApplyUpstreamTLSSettings(t *testing.T) {
-	istioMutualTLSSettingsWithCerts := &networking.ClientTLSSettings{
-		Mode:              networking.ClientTLSSettings_ISTIO_MUTUAL,
-		CaCertificates:    constants.DefaultRootCert,
-		ClientCertificate: constants.DefaultCertChain,
-		PrivateKey:        constants.DefaultKey,
-		SubjectAltNames:   []string{"custom.foo.com"},
-		Sni:               "custom.foo.com",
-	}
-	istioMutualTLSSettings := &networking.ClientTLSSettings{
-		Mode:            networking.ClientTLSSettings_ISTIO_MUTUAL,
-		SubjectAltNames: []string{"custom.foo.com"},
-		Sni:             "custom.foo.com",
-	}
-	mutualTLSSettingsWithCerts := &networking.ClientTLSSettings{
-		Mode:              networking.ClientTLSSettings_MUTUAL,
-		CaCertificates:    constants.DefaultRootCert,
-		ClientCertificate: constants.DefaultCertChain,
-		PrivateKey:        constants.DefaultKey,
-		SubjectAltNames:   []string{"custom.foo.com"},
-		Sni:               "custom.foo.com",
-	}
-	simpleTLSSettingsWithCerts := &networking.ClientTLSSettings{
-		Mode:            networking.ClientTLSSettings_SIMPLE,
-		CaCertificates:  constants.DefaultRootCert,
-		SubjectAltNames: []string{"custom.foo.com"},
-		Sni:             "custom.foo.com",
-	}
-
-	http2ProtocolOptions := &core.Http2ProtocolOptions{
-		AllowConnect:  true,
-		AllowMetadata: true,
-	}
-
-	expectedNodeMetadataClientKeyPath := "/clientKeyFromNodeMetadata.pem"
-	expectedNodeMetadataClientCertPath := "/clientCertFromNodeMetadata.pem"
-	expectedNodeMetadataRootCertPath := "/clientRootCertFromNodeMetadata.pem"
-
-	tests := []struct {
-		name                       string
-		mtlsCtx                    mtlsContextType
-		discoveryType              cluster.Cluster_DiscoveryType
-		tls                        *networking.ClientTLSSettings
-		customMetadata             *model.NodeMetadata
-		allowCustomMetadataMutual  bool
-		expectTransportSocket      bool
-		expectTransportSocketMatch bool
-		http2ProtocolOptions       *core.Http2ProtocolOptions
-
-		validateTLSContext func(t *testing.T, ctx *tls.UpstreamTlsContext)
-	}{
-		{
-			name:                       "user specified without tls",
-			mtlsCtx:                    userSupplied,
-			discoveryType:              cluster.Cluster_EDS,
-			tls:                        nil,
-			expectTransportSocket:      false,
-			expectTransportSocketMatch: false,
-		},
-		{
-			name:                       "user specified with istio_mutual metadata certs tls",
-			mtlsCtx:                    userSupplied,
-			discoveryType:              cluster.Cluster_EDS,
-			tls:                        istioMutualTLSSettingsWithCerts,
-			expectTransportSocket:      true,
-			expectTransportSocketMatch: false,
-			validateTLSContext: func(t *testing.T, ctx *tls.UpstreamTlsContext) {
-				if got := ctx.CommonTlsContext.GetAlpnProtocols(); len(got) != 0 {
-					t.Fatalf("expected empty alpn list got %v", got)
-				}
-			},
-		},
-		{
-			name:                       "user specified with istio_mutual metadata certs tls with h2",
-			mtlsCtx:                    userSupplied,
-			discoveryType:              cluster.Cluster_EDS,
-			tls:                        istioMutualTLSSettingsWithCerts,
-			expectTransportSocket:      true,
-			expectTransportSocketMatch: false,
-			http2ProtocolOptions:       http2ProtocolOptions,
-			validateTLSContext: func(t *testing.T, ctx *tls.UpstreamTlsContext) {
-				if got := ctx.CommonTlsContext.GetAlpnProtocols(); !reflect.DeepEqual(got, util.ALPNH2Only) {
-					t.Fatalf("expected alpn list %v; got %v", util.ALPNH2Only, got)
-				}
-			},
-		},
-		{
-			name:                       "user specified with istio_mutual tls",
-			mtlsCtx:                    userSupplied,
-			discoveryType:              cluster.Cluster_EDS,
-			tls:                        istioMutualTLSSettings,
-			expectTransportSocket:      true,
-			expectTransportSocketMatch: false,
-			validateTLSContext: func(t *testing.T, ctx *tls.UpstreamTlsContext) {
-				if got := ctx.CommonTlsContext.GetAlpnProtocols(); !reflect.DeepEqual(got, util.ALPNInMeshWithMxc) {
-					t.Fatalf("expected alpn list %v; got %v", util.ALPNInMeshWithMxc, got)
-				}
-			},
-		},
-		{
-			name:                       "user specified with istio_mutual tls with h2",
-			mtlsCtx:                    userSupplied,
-			discoveryType:              cluster.Cluster_EDS,
-			tls:                        istioMutualTLSSettings,
-			expectTransportSocket:      true,
-			expectTransportSocketMatch: false,
-			http2ProtocolOptions:       http2ProtocolOptions,
-			validateTLSContext: func(t *testing.T, ctx *tls.UpstreamTlsContext) {
-				if got := ctx.CommonTlsContext.GetAlpnProtocols(); !reflect.DeepEqual(got, util.ALPNInMeshH2WithMxc) {
-					t.Fatalf("expected alpn list %v; got %v", util.ALPNInMeshH2WithMxc, got)
-				}
-			},
-		},
-		{
-			name:                       "user specified simple tls",
-			mtlsCtx:                    userSupplied,
-			discoveryType:              cluster.Cluster_EDS,
-			tls:                        simpleTLSSettingsWithCerts,
-			expectTransportSocket:      true,
-			expectTransportSocketMatch: false,
-			validateTLSContext: func(t *testing.T, ctx *tls.UpstreamTlsContext) {
-				rootName := "file-root:" + mutualTLSSettingsWithCerts.CaCertificates
-				if got := ctx.CommonTlsContext.GetCombinedValidationContext().GetValidationContextSdsSecretConfig().GetName(); rootName != got {
-					t.Fatalf("expected root name %v got %v", rootName, got)
-				}
-				if got := ctx.CommonTlsContext.GetAlpnProtocols(); got != nil {
-					t.Fatalf("expected alpn list nil as not h2 or Istio_Mutual TLS Setting; got %v", got)
-				}
-				if got := ctx.GetSni(); got != simpleTLSSettingsWithCerts.Sni {
-					t.Fatalf("expected TLSContext SNI %v; got %v", simpleTLSSettingsWithCerts.Sni, got)
-				}
-			},
-		},
-		{
-			name:                       "user specified simple tls with h2",
-			mtlsCtx:                    userSupplied,
-			discoveryType:              cluster.Cluster_EDS,
-			tls:                        simpleTLSSettingsWithCerts,
-			expectTransportSocket:      true,
-			expectTransportSocketMatch: false,
-			http2ProtocolOptions:       http2ProtocolOptions,
-			validateTLSContext: func(t *testing.T, ctx *tls.UpstreamTlsContext) {
-				rootName := "file-root:" + mutualTLSSettingsWithCerts.CaCertificates
-				if got := ctx.CommonTlsContext.GetCombinedValidationContext().GetValidationContextSdsSecretConfig().GetName(); rootName != got {
-					t.Fatalf("expected root name %v got %v", rootName, got)
-				}
-				if got := ctx.CommonTlsContext.GetAlpnProtocols(); !reflect.DeepEqual(got, util.ALPNH2Only) {
-					t.Fatalf("expected alpn list %v; got %v", util.ALPNH2Only, got)
-				}
-				if got := ctx.GetSni(); got != simpleTLSSettingsWithCerts.Sni {
-					t.Fatalf("expected TLSContext SNI %v; got %v", simpleTLSSettingsWithCerts.Sni, got)
-				}
-			},
-		},
-		{
-			name:                       "user specified mutual tls",
-			mtlsCtx:                    userSupplied,
-			discoveryType:              cluster.Cluster_EDS,
-			tls:                        mutualTLSSettingsWithCerts,
-			expectTransportSocket:      true,
-			expectTransportSocketMatch: false,
-			validateTLSContext: func(t *testing.T, ctx *tls.UpstreamTlsContext) {
-				rootName := "file-root:" + mutualTLSSettingsWithCerts.CaCertificates
-				certName := fmt.Sprintf("file-cert:%s~%s", mutualTLSSettingsWithCerts.ClientCertificate, mutualTLSSettingsWithCerts.PrivateKey)
-				if got := ctx.CommonTlsContext.GetCombinedValidationContext().GetValidationContextSdsSecretConfig().GetName(); rootName != got {
-					t.Fatalf("expected root name %v got %v", rootName, got)
-				}
-				if got := ctx.CommonTlsContext.GetTlsCertificateSdsSecretConfigs()[0].GetName(); certName != got {
-					t.Fatalf("expected cert name %v got %v", certName, got)
-				}
-				if got := ctx.CommonTlsContext.GetAlpnProtocols(); got != nil {
-					t.Fatalf("expected alpn list nil as not h2 or Istio_Mutual TLS Setting; got %v", got)
-				}
-				if got := ctx.GetSni(); got != mutualTLSSettingsWithCerts.Sni {
-					t.Fatalf("expected TLSContext SNI %v; got %v", mutualTLSSettingsWithCerts.Sni, got)
-				}
-			},
-		},
-		{
-			name:                       "user specified mutual tls with h2",
-			mtlsCtx:                    userSupplied,
-			discoveryType:              cluster.Cluster_EDS,
-			tls:                        mutualTLSSettingsWithCerts,
-			expectTransportSocket:      true,
-			expectTransportSocketMatch: false,
-			http2ProtocolOptions:       http2ProtocolOptions,
-			validateTLSContext: func(t *testing.T, ctx *tls.UpstreamTlsContext) {
-				rootName := "file-root:" + mutualTLSSettingsWithCerts.CaCertificates
-				certName := fmt.Sprintf("file-cert:%s~%s", mutualTLSSettingsWithCerts.ClientCertificate, mutualTLSSettingsWithCerts.PrivateKey)
-				if got := ctx.CommonTlsContext.GetCombinedValidationContext().GetValidationContextSdsSecretConfig().GetName(); rootName != got {
-					t.Fatalf("expected root name %v got %v", rootName, got)
-				}
-				if got := ctx.CommonTlsContext.GetTlsCertificateSdsSecretConfigs()[0].GetName(); certName != got {
-					t.Fatalf("expected cert name %v got %v", certName, got)
-				}
-				if got := ctx.CommonTlsContext.GetAlpnProtocols(); !reflect.DeepEqual(got, util.ALPNH2Only) {
-					t.Fatalf("expected alpn list %v; got %v", util.ALPNH2Only, got)
-				}
-				if got := ctx.GetSni(); got != mutualTLSSettingsWithCerts.Sni {
-					t.Fatalf("expected TLSContext SNI %v; got %v", mutualTLSSettingsWithCerts.Sni, got)
-				}
-			},
-		},
-		{
-			name:                       "auto detect with tls",
-			mtlsCtx:                    autoDetected,
-			discoveryType:              cluster.Cluster_EDS,
-			tls:                        istioMutualTLSSettings,
-			expectTransportSocket:      false,
-			expectTransportSocketMatch: true,
-			validateTLSContext: func(t *testing.T, ctx *tls.UpstreamTlsContext) {
-				if got := ctx.CommonTlsContext.GetAlpnProtocols(); !reflect.DeepEqual(got, util.ALPNInMeshWithMxc) {
-					t.Fatalf("expected alpn list %v; got %v", util.ALPNInMeshWithMxc, got)
-				}
-			},
-		},
-		{
-			name:                       "auto detect with tls and h2 options",
-			mtlsCtx:                    autoDetected,
-			discoveryType:              cluster.Cluster_EDS,
-			tls:                        istioMutualTLSSettings,
-			expectTransportSocket:      false,
-			expectTransportSocketMatch: true,
-			http2ProtocolOptions:       http2ProtocolOptions,
-			validateTLSContext: func(t *testing.T, ctx *tls.UpstreamTlsContext) {
-				if got := ctx.CommonTlsContext.GetAlpnProtocols(); !reflect.DeepEqual(got, util.ALPNInMeshH2WithMxc) {
-					t.Fatalf("expected alpn list %v; got %v", util.ALPNInMeshH2WithMxc, got)
-				}
-			},
-		},
-		{
-			name:                       "auto detect with tls",
-			mtlsCtx:                    autoDetected,
-			discoveryType:              cluster.Cluster_ORIGINAL_DST,
-			tls:                        istioMutualTLSSettingsWithCerts,
-			expectTransportSocket:      true,
-			expectTransportSocketMatch: false,
-		},
-		{
-			name:          "user specified mutual tls with overridden certs from node metadata allowed",
-			mtlsCtx:       userSupplied,
-			discoveryType: cluster.Cluster_EDS,
-			tls:           mutualTLSSettingsWithCerts,
-			customMetadata: &model.NodeMetadata{
-				TLSClientCertChain: expectedNodeMetadataClientCertPath,
-				TLSClientKey:       expectedNodeMetadataClientKeyPath,
-				TLSClientRootCert:  expectedNodeMetadataRootCertPath,
-			},
-			allowCustomMetadataMutual:  true,
-			expectTransportSocket:      true,
-			expectTransportSocketMatch: false,
-			validateTLSContext: func(t *testing.T, ctx *tls.UpstreamTlsContext) {
-				rootName := "file-root:" + expectedNodeMetadataRootCertPath
-				certName := fmt.Sprintf("file-cert:%s~%s", expectedNodeMetadataClientCertPath, expectedNodeMetadataClientKeyPath)
-				if got := ctx.CommonTlsContext.GetCombinedValidationContext().GetValidationContextSdsSecretConfig().GetName(); rootName != got {
-					t.Fatalf("expected root name %v got %v", rootName, got)
-				}
-				if got := ctx.CommonTlsContext.GetTlsCertificateSdsSecretConfigs()[0].GetName(); certName != got {
-					t.Fatalf("expected cert name %v got %v", certName, got)
-				}
-				if got := ctx.CommonTlsContext.GetAlpnProtocols(); got != nil {
-					t.Fatalf("expected alpn list nil as not h2 or Istio_Mutual TLS Setting; got %v", got)
-				}
-			},
-		},
-		{
-			name:          "user specified mutual tls with overridden certs from node metadata not allowed",
-			mtlsCtx:       userSupplied,
-			discoveryType: cluster.Cluster_EDS,
-			tls:           mutualTLSSettingsWithCerts,
-			customMetadata: &model.NodeMetadata{
-				TLSClientCertChain: expectedNodeMetadataClientCertPath,
-				TLSClientKey:       expectedNodeMetadataClientKeyPath,
-				TLSClientRootCert:  expectedNodeMetadataRootCertPath,
-			},
-			expectTransportSocket:      true,
-			expectTransportSocketMatch: false,
-			validateTLSContext: func(t *testing.T, ctx *tls.UpstreamTlsContext) {
-				rootName := "file-root:" + mutualTLSSettingsWithCerts.CaCertificates
-				certName := fmt.Sprintf("file-cert:%s~%s", mutualTLSSettingsWithCerts.ClientCertificate, mutualTLSSettingsWithCerts.PrivateKey)
-				if got := ctx.CommonTlsContext.GetCombinedValidationContext().GetValidationContextSdsSecretConfig().GetName(); rootName != got {
-					t.Fatalf("expected root name %v got %v", rootName, got)
-				}
-				if got := ctx.CommonTlsContext.GetTlsCertificateSdsSecretConfigs()[0].GetName(); certName != got {
-					t.Fatalf("expected cert name %v got %v", certName, got)
-				}
-				if got := ctx.CommonTlsContext.GetAlpnProtocols(); got != nil {
-					t.Fatalf("expected alpn list nil as not h2 or Istio_Mutual TLS Setting; got %v", got)
-				}
-			},
-		},
-	}
-
-	proxy := &model.Proxy{
-		Type:         model.SidecarProxy,
-		Metadata:     &model.NodeMetadata{},
-		IstioVersion: &model.IstioVersion{Major: 1, Minor: 5},
-	}
-	push := model.NewPushContext()
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			cb := NewClusterBuilder(proxy, push)
-			customMetadataMutual := features.AllowMetadataCertsInMutualTLS
-			if test.allowCustomMetadataMutual {
-				features.AllowMetadataCertsInMutualTLS = true
-			}
-			defer func() { features.AllowMetadataCertsInMutualTLS = customMetadataMutual }()
-			if test.customMetadata != nil {
-				proxy.Metadata = test.customMetadata
-			} else {
-				proxy.Metadata = &model.NodeMetadata{}
-			}
-			opts := &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					ClusterDiscoveryType: &cluster.Cluster_Type{Type: test.discoveryType},
-					Http2ProtocolOptions: test.http2ProtocolOptions,
-				},
-				proxy: proxy,
-				mesh:  push.Mesh,
-			}
-			cb.applyUpstreamTLSSettings(opts, test.tls, test.mtlsCtx)
-
-			if test.expectTransportSocket && opts.cluster.TransportSocket == nil ||
-				!test.expectTransportSocket && opts.cluster.TransportSocket != nil {
-				t.Errorf("Expected TransportSocket %v", test.expectTransportSocket)
-			}
-			if test.expectTransportSocketMatch && opts.cluster.TransportSocketMatches == nil ||
-				!test.expectTransportSocketMatch && opts.cluster.TransportSocketMatches != nil {
-				t.Errorf("Expected TransportSocketMatch %v", test.expectTransportSocketMatch)
-			}
-
-			if test.validateTLSContext != nil {
-				ctx := &tls.UpstreamTlsContext{}
-				if test.expectTransportSocket {
-					if err := ptypes.UnmarshalAny(opts.cluster.TransportSocket.GetTypedConfig(), ctx); err != nil {
-						t.Fatal(err)
-					}
-				} else if test.expectTransportSocketMatch {
-					if err := ptypes.UnmarshalAny(opts.cluster.TransportSocketMatches[0].TransportSocket.GetTypedConfig(), ctx); err != nil {
-						t.Fatal(err)
-					}
-				}
-				test.validateTLSContext(t, ctx)
-			}
-		})
-	}
-}
-
-type expectedResult struct {
-	tlsContext *tls.UpstreamTlsContext
-	err        error
-}
-
-// TestBuildUpstreamClusterTLSContext tests the buildUpstreamClusterTLSContext function
-func TestBuildUpstreamClusterTLSContext(t *testing.T) {
-	metadataClientCert := "/path/to/metadata/cert"
-	metadataRootCert := "/path/to/metadata/root-cert"
-	metadataClientKey := "/path/to/metadata/key"
-
-	clientCert := "/path/to/cert"
-	rootCert := "path/to/cacert"
-	clientKey := "/path/to/key"
-
-	credentialName := "some-fake-credential"
-
-	testCases := []struct {
-		name   string
-		opts   *buildClusterOpts
-		tls    *networking.ClientTLSSettings
-		result expectedResult
-	}{
-		{
-			name: "tls mode disabled",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode: networking.ClientTLSSettings_DISABLE,
-			},
-			result: expectedResult{nil, nil},
-		},
-		{
-			name: "tls mode ISTIO_MUTUAL with metadata certs",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:              networking.ClientTLSSettings_ISTIO_MUTUAL,
-				ClientCertificate: metadataClientCert,
-				PrivateKey:        metadataClientKey,
-				CaCertificates:    metadataRootCert,
-				SubjectAltNames:   []string{"SAN"},
-				Sni:               "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						TlsCertificateSdsSecretConfigs: []*tls.SdsSecretConfig{
-							{
-								Name: "file-cert:/path/to/metadata/cert~/path/to/metadata/key",
-								SdsConfig: &core.ConfigSource{
-									ConfigSourceSpecifier: &core.ConfigSource_ApiConfigSource{
-										ApiConfigSource: &core.ApiConfigSource{
-											ApiType:                   core.ApiConfigSource_GRPC,
-											SetNodeOnFirstMessageOnly: true,
-											TransportApiVersion:       core.ApiVersion_V3,
-											GrpcServices: []*core.GrpcService{
-												{
-													TargetSpecifier: &core.GrpcService_EnvoyGrpc_{
-														EnvoyGrpc: &core.GrpcService_EnvoyGrpc{ClusterName: "sds-grpc"},
-													},
-												},
-											},
-										},
-									},
-									ResourceApiVersion: core.ApiVersion_V3,
-								},
-							},
-						},
-						ValidationContextType: &tls.CommonTlsContext_CombinedValidationContext{
-							CombinedValidationContext: &tls.CommonTlsContext_CombinedCertificateValidationContext{
-								DefaultValidationContext: &tls.CertificateValidationContext{MatchSubjectAltNames: util.StringToExactMatch([]string{"SAN"})},
-								ValidationContextSdsSecretConfig: &tls.SdsSecretConfig{
-									Name: "file-root:/path/to/metadata/root-cert",
-									SdsConfig: &core.ConfigSource{
-										ConfigSourceSpecifier: &core.ConfigSource_ApiConfigSource{
-											ApiConfigSource: &core.ApiConfigSource{
-												ApiType:                   core.ApiConfigSource_GRPC,
-												SetNodeOnFirstMessageOnly: true,
-												TransportApiVersion:       core.ApiVersion_V3,
-												GrpcServices: []*core.GrpcService{
-													{
-														TargetSpecifier: &core.GrpcService_EnvoyGrpc_{
-															EnvoyGrpc: &core.GrpcService_EnvoyGrpc{ClusterName: "sds-grpc"},
-														},
-													},
-												},
-											},
-										},
-										ResourceApiVersion: core.ApiVersion_V3,
-									},
-								},
-							},
-						},
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode ISTIO_MUTUAL with metadata certs and H2",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name:                 "test-cluster",
-					Http2ProtocolOptions: &core.Http2ProtocolOptions{},
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:              networking.ClientTLSSettings_ISTIO_MUTUAL,
-				ClientCertificate: metadataClientCert,
-				PrivateKey:        metadataClientKey,
-				CaCertificates:    metadataRootCert,
-				SubjectAltNames:   []string{"SAN"},
-				Sni:               "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						TlsCertificateSdsSecretConfigs: []*tls.SdsSecretConfig{
-							{
-								Name: "file-cert:/path/to/metadata/cert~/path/to/metadata/key",
-								SdsConfig: &core.ConfigSource{
-									ConfigSourceSpecifier: &core.ConfigSource_ApiConfigSource{
-										ApiConfigSource: &core.ApiConfigSource{
-											ApiType:                   core.ApiConfigSource_GRPC,
-											SetNodeOnFirstMessageOnly: true,
-											TransportApiVersion:       core.ApiVersion_V3,
-											GrpcServices: []*core.GrpcService{
-												{
-													TargetSpecifier: &core.GrpcService_EnvoyGrpc_{
-														EnvoyGrpc: &core.GrpcService_EnvoyGrpc{ClusterName: "sds-grpc"},
-													},
-												},
-											},
-										},
-									},
-									ResourceApiVersion: core.ApiVersion_V3,
-								},
-							},
-						},
-						ValidationContextType: &tls.CommonTlsContext_CombinedValidationContext{
-							CombinedValidationContext: &tls.CommonTlsContext_CombinedCertificateValidationContext{
-								DefaultValidationContext: &tls.CertificateValidationContext{MatchSubjectAltNames: util.StringToExactMatch([]string{"SAN"})},
-								ValidationContextSdsSecretConfig: &tls.SdsSecretConfig{
-									Name: "file-root:/path/to/metadata/root-cert",
-									SdsConfig: &core.ConfigSource{
-										ConfigSourceSpecifier: &core.ConfigSource_ApiConfigSource{
-											ApiConfigSource: &core.ApiConfigSource{
-												ApiType:                   core.ApiConfigSource_GRPC,
-												SetNodeOnFirstMessageOnly: true,
-												TransportApiVersion:       core.ApiVersion_V3,
-												GrpcServices: []*core.GrpcService{
-													{
-														TargetSpecifier: &core.GrpcService_EnvoyGrpc_{
-															EnvoyGrpc: &core.GrpcService_EnvoyGrpc{ClusterName: "sds-grpc"},
-														},
-													},
-												},
-											},
-										},
-										ResourceApiVersion: core.ApiVersion_V3,
-									},
-								},
-							},
-						},
-						AlpnProtocols: util.ALPNH2Only,
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode SIMPLE, with no certs specified in tls",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:            networking.ClientTLSSettings_SIMPLE,
-				SubjectAltNames: []string{"SAN"},
-				Sni:             "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						ValidationContextType: &tls.CommonTlsContext_ValidationContext{},
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode SIMPLE, with certs specified in tls",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:            networking.ClientTLSSettings_SIMPLE,
-				CaCertificates:  rootCert,
-				SubjectAltNames: []string{"SAN"},
-				Sni:             "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						ValidationContextType: &tls.CommonTlsContext_CombinedValidationContext{
-							CombinedValidationContext: &tls.CommonTlsContext_CombinedCertificateValidationContext{
-								DefaultValidationContext: &tls.CertificateValidationContext{MatchSubjectAltNames: util.StringToExactMatch([]string{"SAN"})},
-								ValidationContextSdsSecretConfig: &tls.SdsSecretConfig{
-									Name: fmt.Sprintf("file-root:%s", rootCert),
-									SdsConfig: &core.ConfigSource{
-										ConfigSourceSpecifier: &core.ConfigSource_ApiConfigSource{
-											ApiConfigSource: &core.ApiConfigSource{
-												ApiType:                   core.ApiConfigSource_GRPC,
-												SetNodeOnFirstMessageOnly: true,
-												TransportApiVersion:       core.ApiVersion_V3,
-												GrpcServices: []*core.GrpcService{
-													{
-														TargetSpecifier: &core.GrpcService_EnvoyGrpc_{
-															EnvoyGrpc: &core.GrpcService_EnvoyGrpc{ClusterName: "sds-grpc"},
-														},
-													},
-												},
-											},
-										},
-										ResourceApiVersion: core.ApiVersion_V3,
-									},
-								},
-							},
-						},
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode SIMPLE, with certs specified in tls with h2",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name:                 "test-cluster",
-					Http2ProtocolOptions: &core.Http2ProtocolOptions{},
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:            networking.ClientTLSSettings_SIMPLE,
-				CaCertificates:  rootCert,
-				SubjectAltNames: []string{"SAN"},
-				Sni:             "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						ValidationContextType: &tls.CommonTlsContext_CombinedValidationContext{
-							CombinedValidationContext: &tls.CommonTlsContext_CombinedCertificateValidationContext{
-								DefaultValidationContext: &tls.CertificateValidationContext{MatchSubjectAltNames: util.StringToExactMatch([]string{"SAN"})},
-								ValidationContextSdsSecretConfig: &tls.SdsSecretConfig{
-									Name: fmt.Sprintf("file-root:%s", rootCert),
-									SdsConfig: &core.ConfigSource{
-										ConfigSourceSpecifier: &core.ConfigSource_ApiConfigSource{
-											ApiConfigSource: &core.ApiConfigSource{
-												ApiType:                   core.ApiConfigSource_GRPC,
-												SetNodeOnFirstMessageOnly: true,
-												TransportApiVersion:       core.ApiVersion_V3,
-												GrpcServices: []*core.GrpcService{
-													{
-														TargetSpecifier: &core.GrpcService_EnvoyGrpc_{
-															EnvoyGrpc: &core.GrpcService_EnvoyGrpc{ClusterName: "sds-grpc"},
-														},
-													},
-												},
-											},
-										},
-										ResourceApiVersion: core.ApiVersion_V3,
-									},
-								},
-							},
-						},
-						AlpnProtocols: util.ALPNH2Only,
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode SIMPLE, with certs specified in tls with overridden metadata certs",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{
-						TLSClientRootCert: metadataRootCert,
-					},
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:            networking.ClientTLSSettings_SIMPLE,
-				CaCertificates:  rootCert,
-				SubjectAltNames: []string{"SAN"},
-				Sni:             "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						ValidationContextType: &tls.CommonTlsContext_CombinedValidationContext{
-							CombinedValidationContext: &tls.CommonTlsContext_CombinedCertificateValidationContext{
-								DefaultValidationContext: &tls.CertificateValidationContext{MatchSubjectAltNames: util.StringToExactMatch([]string{"SAN"})},
-								ValidationContextSdsSecretConfig: &tls.SdsSecretConfig{
-									Name: fmt.Sprintf("file-root:%s", metadataRootCert),
-									SdsConfig: &core.ConfigSource{
-										ConfigSourceSpecifier: &core.ConfigSource_ApiConfigSource{
-											ApiConfigSource: &core.ApiConfigSource{
-												ApiType:                   core.ApiConfigSource_GRPC,
-												SetNodeOnFirstMessageOnly: true,
-												TransportApiVersion:       core.ApiVersion_V3,
-												GrpcServices: []*core.GrpcService{
-													{
-														TargetSpecifier: &core.GrpcService_EnvoyGrpc_{
-															EnvoyGrpc: &core.GrpcService_EnvoyGrpc{ClusterName: "sds-grpc"},
-														},
-													},
-												},
-											},
-										},
-										ResourceApiVersion: core.ApiVersion_V3,
-									},
-								},
-							},
-						},
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode MUTUAL, with no client certificate",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:              networking.ClientTLSSettings_MUTUAL,
-				ClientCertificate: "",
-				PrivateKey:        "some-fake-key",
-			},
-			result: expectedResult{
-				nil,
-				fmt.Errorf("client cert must be provided"),
-			},
-		},
-		{
-			name: "tls mode MUTUAL, with no client key",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:              networking.ClientTLSSettings_MUTUAL,
-				ClientCertificate: "some-fake-cert",
-				PrivateKey:        "",
-			},
-			result: expectedResult{
-				nil,
-				fmt.Errorf("client key must be provided"),
-			},
-		},
-		{
-			name: "tls mode MUTUAL, with node metadata sdsEnabled true no root CA specified",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:              networking.ClientTLSSettings_MUTUAL,
-				ClientCertificate: clientCert,
-				PrivateKey:        clientKey,
-				SubjectAltNames:   []string{"SAN"},
-				Sni:               "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						TlsCertificateSdsSecretConfigs: []*tls.SdsSecretConfig{
-							{
-								Name: fmt.Sprintf("file-cert:%s~%s", clientCert, clientKey),
-								SdsConfig: &core.ConfigSource{
-									ConfigSourceSpecifier: &core.ConfigSource_ApiConfigSource{
-										ApiConfigSource: &core.ApiConfigSource{
-											ApiType:                   core.ApiConfigSource_GRPC,
-											SetNodeOnFirstMessageOnly: true,
-											TransportApiVersion:       core.ApiVersion_V3,
-											GrpcServices: []*core.GrpcService{
-												{
-													TargetSpecifier: &core.GrpcService_EnvoyGrpc_{
-														EnvoyGrpc: &core.GrpcService_EnvoyGrpc{ClusterName: "sds-grpc"},
-													},
-												},
-											},
-										},
-									},
-									ResourceApiVersion: core.ApiVersion_V3,
-								},
-							},
-						},
-						ValidationContextType: &tls.CommonTlsContext_ValidationContext{},
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode MUTUAL, with node metadata sdsEnabled true",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:              networking.ClientTLSSettings_MUTUAL,
-				ClientCertificate: clientCert,
-				PrivateKey:        clientKey,
-				CaCertificates:    rootCert,
-				SubjectAltNames:   []string{"SAN"},
-				Sni:               "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						TlsCertificateSdsSecretConfigs: []*tls.SdsSecretConfig{
-							{
-								Name: fmt.Sprintf("file-cert:%s~%s", clientCert, clientKey),
-								SdsConfig: &core.ConfigSource{
-									ConfigSourceSpecifier: &core.ConfigSource_ApiConfigSource{
-										ApiConfigSource: &core.ApiConfigSource{
-											ApiType:                   core.ApiConfigSource_GRPC,
-											SetNodeOnFirstMessageOnly: true,
-											TransportApiVersion:       core.ApiVersion_V3,
-											GrpcServices: []*core.GrpcService{
-												{
-													TargetSpecifier: &core.GrpcService_EnvoyGrpc_{
-														EnvoyGrpc: &core.GrpcService_EnvoyGrpc{ClusterName: "sds-grpc"},
-													},
-												},
-											},
-										},
-									},
-									ResourceApiVersion: core.ApiVersion_V3,
-								},
-							},
-						},
-						ValidationContextType: &tls.CommonTlsContext_CombinedValidationContext{
-							CombinedValidationContext: &tls.CommonTlsContext_CombinedCertificateValidationContext{
-								DefaultValidationContext: &tls.CertificateValidationContext{MatchSubjectAltNames: util.StringToExactMatch([]string{"SAN"})},
-								ValidationContextSdsSecretConfig: &tls.SdsSecretConfig{
-									Name: fmt.Sprintf("file-root:%s", rootCert),
-									SdsConfig: &core.ConfigSource{
-										ConfigSourceSpecifier: &core.ConfigSource_ApiConfigSource{
-											ApiConfigSource: &core.ApiConfigSource{
-												ApiType:                   core.ApiConfigSource_GRPC,
-												SetNodeOnFirstMessageOnly: true,
-												TransportApiVersion:       core.ApiVersion_V3,
-												GrpcServices: []*core.GrpcService{
-													{
-														TargetSpecifier: &core.GrpcService_EnvoyGrpc_{
-															EnvoyGrpc: &core.GrpcService_EnvoyGrpc{ClusterName: "sds-grpc"},
-														},
-													},
-												},
-											},
-										},
-										ResourceApiVersion: core.ApiVersion_V3,
-									},
-								},
-							},
-						},
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode SIMPLE, with CredentialName specified",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-					Type:     model.Router,
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:            networking.ClientTLSSettings_SIMPLE,
-				CredentialName:  credentialName,
-				SubjectAltNames: []string{"SAN"},
-				Sni:             "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						ValidationContextType: &tls.CommonTlsContext_CombinedValidationContext{
-							CombinedValidationContext: &tls.CommonTlsContext_CombinedCertificateValidationContext{
-								DefaultValidationContext: &tls.CertificateValidationContext{
-									MatchSubjectAltNames: util.StringToExactMatch([]string{"SAN"}),
-								},
-								ValidationContextSdsSecretConfig: &tls.SdsSecretConfig{
-									Name:      "kubernetes://" + credentialName + authn_model.SdsCaSuffix,
-									SdsConfig: authn_model.SDSAdsConfig,
-								},
-							},
-						},
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode SIMPLE, with CredentialName specified",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-					Type:     model.Router,
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:            networking.ClientTLSSettings_SIMPLE,
-				CredentialName:  credentialName,
-				SubjectAltNames: []string{"SAN"},
-				Sni:             "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						ValidationContextType: &tls.CommonTlsContext_CombinedValidationContext{
-							CombinedValidationContext: &tls.CommonTlsContext_CombinedCertificateValidationContext{
-								DefaultValidationContext: &tls.CertificateValidationContext{
-									MatchSubjectAltNames: util.StringToExactMatch([]string{"SAN"}),
-								},
-								ValidationContextSdsSecretConfig: &tls.SdsSecretConfig{
-									Name:      "kubernetes://" + credentialName + authn_model.SdsCaSuffix,
-									SdsConfig: authn_model.SDSAdsConfig,
-								},
-							},
-						},
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode SIMPLE, with CredentialName specified with h2 and no SAN",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name:                 "test-cluster",
-					Http2ProtocolOptions: &core.Http2ProtocolOptions{},
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-					Type:     model.Router,
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:           networking.ClientTLSSettings_SIMPLE,
-				CredentialName: credentialName,
-				Sni:            "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						ValidationContextType: &tls.CommonTlsContext_CombinedValidationContext{
-							CombinedValidationContext: &tls.CommonTlsContext_CombinedCertificateValidationContext{
-								DefaultValidationContext: &tls.CertificateValidationContext{},
-								ValidationContextSdsSecretConfig: &tls.SdsSecretConfig{
-									Name:      "kubernetes://" + credentialName + authn_model.SdsCaSuffix,
-									SdsConfig: authn_model.SDSAdsConfig,
-								},
-							},
-						},
-						AlpnProtocols: util.ALPNH2Only,
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode MUTUAL, with CredentialName specified",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-					Type:     model.Router,
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:            networking.ClientTLSSettings_MUTUAL,
-				CredentialName:  credentialName,
-				SubjectAltNames: []string{"SAN"},
-				Sni:             "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						TlsCertificateSdsSecretConfigs: []*tls.SdsSecretConfig{
-							{
-								Name:      "kubernetes://" + credentialName,
-								SdsConfig: authn_model.SDSAdsConfig,
-							},
-						},
-						ValidationContextType: &tls.CommonTlsContext_CombinedValidationContext{
-							CombinedValidationContext: &tls.CommonTlsContext_CombinedCertificateValidationContext{
-								DefaultValidationContext: &tls.CertificateValidationContext{
-									MatchSubjectAltNames: util.StringToExactMatch([]string{"SAN"}),
-								},
-								ValidationContextSdsSecretConfig: &tls.SdsSecretConfig{
-									Name:      "kubernetes://" + credentialName + authn_model.SdsCaSuffix,
-									SdsConfig: authn_model.SDSAdsConfig,
-								},
-							},
-						},
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode MUTUAL, with CredentialName specified with h2 and no SAN",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name:                 "test-cluster",
-					Http2ProtocolOptions: &core.Http2ProtocolOptions{},
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-					Type:     model.Router,
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:           networking.ClientTLSSettings_MUTUAL,
-				CredentialName: credentialName,
-				Sni:            "some-sni.com",
-			},
-			result: expectedResult{
-				tlsContext: &tls.UpstreamTlsContext{
-					CommonTlsContext: &tls.CommonTlsContext{
-						TlsCertificateSdsSecretConfigs: []*tls.SdsSecretConfig{
-							{
-								Name:      "kubernetes://" + credentialName,
-								SdsConfig: authn_model.SDSAdsConfig,
-							},
-						},
-						ValidationContextType: &tls.CommonTlsContext_CombinedValidationContext{
-							CombinedValidationContext: &tls.CommonTlsContext_CombinedCertificateValidationContext{
-								DefaultValidationContext: &tls.CertificateValidationContext{},
-								ValidationContextSdsSecretConfig: &tls.SdsSecretConfig{
-									Name:      "kubernetes://" + credentialName + authn_model.SdsCaSuffix,
-									SdsConfig: authn_model.SDSAdsConfig,
-								},
-							},
-						},
-						AlpnProtocols: util.ALPNH2Only,
-					},
-					Sni: "some-sni.com",
-				},
-				err: nil,
-			},
-		},
-		{
-			name: "tls mode MUTUAL, credentialName is set with proxy type Sidecar",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-					Type:     model.SidecarProxy,
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:           networking.ClientTLSSettings_MUTUAL,
-				CredentialName: "fake-cred",
-			},
-			result: expectedResult{
-				nil,
-				nil,
-			},
-		},
-		{
-			name: "tls mode SIMPLE, credentialName is set with proxy type Sidecar",
-			opts: &buildClusterOpts{
-				cluster: &cluster.Cluster{
-					Name: "test-cluster",
-				},
-				proxy: &model.Proxy{
-					Metadata: &model.NodeMetadata{},
-					Type:     model.SidecarProxy,
-				},
-			},
-			tls: &networking.ClientTLSSettings{
-				Mode:           networking.ClientTLSSettings_SIMPLE,
-				CredentialName: "fake-cred",
-			},
-			result: expectedResult{
-				nil,
-				nil,
-			},
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			cb := NewClusterBuilder(tc.opts.proxy, nil)
-			ret, err := cb.buildUpstreamClusterTLSContext(tc.opts, tc.tls)
-			if err != nil && tc.result.err == nil || err == nil && tc.result.err != nil {
-				t.Errorf("expecting:\n err=%v but got err=%v", tc.result.err, err)
-			} else if diff := cmp.Diff(tc.result.tlsContext, ret, protocmp.Transform()); diff != "" {
-				t.Errorf("got diff: `%v", diff)
-			}
-		})
-	}
-}
-
-// Helper function to extract TLS context from a cluster
-func getTLSContext(t *testing.T, c *cluster.Cluster) *tls.UpstreamTlsContext {
-	t.Helper()
-	if c.TransportSocket == nil {
-		return nil
-	}
-	tlsContext := &tls.UpstreamTlsContext{}
-	err := ptypes.UnmarshalAny(c.TransportSocket.GetTypedConfig(), tlsContext)
-	if err != nil {
-		t.Fatalf("Failed to unmarshall tls context: %v", err)
-	}
-	return tlsContext
-}
-
 func TestBuildStaticClusterWithNoEndPoint(t *testing.T) {
 	g := NewWithT(t)
 
@@ -3030,123 +1895,6 @@ func TestBuildStaticClusterWithNoEndPoint(t *testing.T) {
 
 	// Expect to ignore STRICT_DNS cluster without endpoints.
 	g.Expect(xdstest.MapKeys(xdstest.ExtractClusters(clusters))).To(Equal([]string{"BlackHoleCluster", "InboundPassthroughClusterIpv4", "PassthroughCluster"}))
-}
-
-func TestShouldH2Upgrade(t *testing.T) {
-	tests := []struct {
-		name           string
-		clusterName    string
-		direction      model.TrafficDirection
-		port           model.Port
-		mesh           meshconfig.MeshConfig
-		connectionPool networking.ConnectionPoolSettings
-
-		upgrade bool
-	}{
-		{
-			name:        "mesh upgrade - dr default",
-			clusterName: "bar",
-			direction:   model.TrafficDirectionOutbound,
-			port:        model.Port{Protocol: protocol.HTTP},
-			mesh:        meshconfig.MeshConfig{H2UpgradePolicy: meshconfig.MeshConfig_UPGRADE},
-			connectionPool: networking.ConnectionPoolSettings{
-				Http: &networking.ConnectionPoolSettings_HTTPSettings{
-					H2UpgradePolicy: networking.ConnectionPoolSettings_HTTPSettings_DEFAULT,
-				},
-			},
-			upgrade: true,
-		},
-		{
-			name:        "mesh default - dr upgrade non http port",
-			clusterName: "bar",
-			direction:   model.TrafficDirectionOutbound,
-			port:        model.Port{Protocol: protocol.Unsupported},
-			mesh:        meshconfig.MeshConfig{},
-			connectionPool: networking.ConnectionPoolSettings{
-				Http: &networking.ConnectionPoolSettings_HTTPSettings{
-					H2UpgradePolicy: networking.ConnectionPoolSettings_HTTPSettings_UPGRADE,
-				},
-			},
-			upgrade: true,
-		},
-		{
-			name:        "mesh no_upgrade - dr default",
-			clusterName: "bar",
-			direction:   model.TrafficDirectionOutbound,
-			port:        model.Port{Protocol: protocol.HTTP},
-			mesh:        meshconfig.MeshConfig{H2UpgradePolicy: meshconfig.MeshConfig_DO_NOT_UPGRADE},
-			connectionPool: networking.ConnectionPoolSettings{
-				Http: &networking.ConnectionPoolSettings_HTTPSettings{
-					H2UpgradePolicy: networking.ConnectionPoolSettings_HTTPSettings_DEFAULT,
-				},
-			},
-			upgrade: false,
-		},
-		{
-			name:        "mesh no_upgrade - dr upgrade",
-			clusterName: "bar",
-			direction:   model.TrafficDirectionOutbound,
-			port:        model.Port{Protocol: protocol.HTTP},
-			mesh:        meshconfig.MeshConfig{H2UpgradePolicy: meshconfig.MeshConfig_DO_NOT_UPGRADE},
-			connectionPool: networking.ConnectionPoolSettings{
-				Http: &networking.ConnectionPoolSettings_HTTPSettings{
-					H2UpgradePolicy: networking.ConnectionPoolSettings_HTTPSettings_UPGRADE,
-				},
-			},
-			upgrade: true,
-		},
-		{
-			name:        "mesh upgrade - dr no_upgrade",
-			clusterName: "bar",
-			direction:   model.TrafficDirectionOutbound,
-			port:        model.Port{Protocol: protocol.HTTP},
-			mesh:        meshconfig.MeshConfig{H2UpgradePolicy: meshconfig.MeshConfig_UPGRADE},
-			connectionPool: networking.ConnectionPoolSettings{
-				Http: &networking.ConnectionPoolSettings_HTTPSettings{
-					H2UpgradePolicy: networking.ConnectionPoolSettings_HTTPSettings_DO_NOT_UPGRADE,
-				},
-			},
-			upgrade: false,
-		},
-		{
-			name:        "inbound ignore",
-			clusterName: "bar",
-			direction:   model.TrafficDirectionInbound,
-			port:        model.Port{Protocol: protocol.HTTP},
-			mesh:        meshconfig.MeshConfig{H2UpgradePolicy: meshconfig.MeshConfig_UPGRADE},
-			connectionPool: networking.ConnectionPoolSettings{
-				Http: &networking.ConnectionPoolSettings_HTTPSettings{
-					H2UpgradePolicy: networking.ConnectionPoolSettings_HTTPSettings_DEFAULT,
-				},
-			},
-			upgrade: false,
-		},
-		{
-			name:        "non-http",
-			clusterName: "bar",
-			direction:   model.TrafficDirectionOutbound,
-			port:        model.Port{Protocol: protocol.Unsupported},
-			mesh:        meshconfig.MeshConfig{H2UpgradePolicy: meshconfig.MeshConfig_UPGRADE},
-			connectionPool: networking.ConnectionPoolSettings{
-				Http: &networking.ConnectionPoolSettings_HTTPSettings{
-					H2UpgradePolicy: networking.ConnectionPoolSettings_HTTPSettings_DEFAULT,
-				},
-			},
-			upgrade: false,
-		},
-	}
-
-	cb := NewClusterBuilder(nil, nil)
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			upgrade := cb.shouldH2Upgrade(test.clusterName, test.direction, &test.port, &test.mesh, &test.connectionPool)
-
-			if upgrade != test.upgrade {
-				t.Fatalf("got: %t, want: %t (%v, %v)", upgrade, test.upgrade, test.mesh.H2UpgradePolicy, test.connectionPool.Http.H2UpgradePolicy)
-			}
-		})
-	}
 }
 
 func TestEnvoyFilterPatching(t *testing.T) {
